重新依赖
