+++ META-INF/MANIFEST.MF
@@ -1,0 +1,35 @@
+Manifest-Version: 1.0
+Class-Path: spring-oxm-4.3.7.RELEASE.jar spring-messaging-4.3.7.RELEAS
+ E.jar slf4j-api-1.7.24.jar swagger-models-1.5.9.jar tomcat-embed-el-8
+ .5.11.jar jackson-core-2.8.7.jar commons-codec-1.10.jar springfox-spi
+ -2.5.0.jar spring-boot-starter-data-redis-1.5.2.RELEASE.jar spring-bo
+ ot-1.5.2.RELEASE.jar springfox-swagger-ui-2.5.0.jar spring-webmvc-4.3
+ .7.RELEASE.jar snappy-java-1.1.2.6.jar hibernate-validator-5.3.4.Fina
+ l.jar c3p0-0.9.5.2.jar spring-boot-starter-web-1.5.2.RELEASE.jar lz4-
+ 1.3.0.jar spring-aop-4.3.7.RELEASE.jar tomcat-embed-websocket-8.5.11.
+ jar jackson-annotations-2.8.0.jar jackson-databind-2.8.7.jar dom4j-1.
+ 6.1.jar log4j-over-slf4j-1.7.24.jar stax-api-1.0.1.jar springfox-swag
+ ger2-2.5.0.jar jboss-logging-3.3.0.Final.jar jedis-2.9.0.jar spring-b
+ oot-starter-logging-1.5.2.RELEASE.jar spring-data-commons-1.13.1.RELE
+ ASE.jar validation-api-1.1.0.Final.jar spring-core-4.3.7.RELEASE.jar 
+ spring-tx-5.0.2.RELEASE.jar druid-1.0.28.jar spring-web-4.3.7.RELEASE
+ .jar spring-boot-starter-1.5.2.RELEASE.jar xml-apis-1.4.01.jar tomcat
+ -embed-core-8.5.11.jar spring-data-redis-1.8.1.RELEASE.jar HikariCP-j
+ ava6-2.3.13.jar poi-ooxml-3.9.jar spring-context-4.3.7.RELEASE.jar sp
+ ring-retry-1.2.0.RELEASE.jar snakeyaml-1.17.jar swagger-annotations-1
+ .5.9.jar kafka-clients-0.10.1.1.jar spring-plugin-core-1.2.0.RELEASE.
+ jar spring-context-support-4.3.7.RELEASE.jar aspectjweaver-1.8.9.jar 
+ jcl-over-slf4j-1.7.24.jar springfox-spring-web-2.5.0.jar spring-kafka
+ -1.1.3.RELEASE.jar jul-to-slf4j-1.7.24.jar spring-data-keyvalue-1.2.1
+ .RELEASE.jar mysql-connector-java-5.1.41.jar xmlbeans-2.3.0.jar logba
+ ck-classic-1.1.11.jar spring-tx-4.3.7.RELEASE.jar springfox-swagger-c
+ ommon-2.5.0.jar poi-3.9.jar commons-pool2-2.4.2.jar quartz-2.3.0.jar 
+ spring-boot-autoconfigure-1.5.2.RELEASE.jar mapstruct-1.0.0.Final.jar
+  classmate-1.3.3.jar poi-ooxml-schemas-3.9.jar spring-beans-4.3.7.REL
+ EASE.jar guava-18.0.jar logback-core-1.1.11.jar springfox-core-2.5.0.
+ jar mchange-commons-java-0.2.11.jar spring-boot-starter-aop-1.5.2.REL
+ EASE.jar spring-expression-4.3.7.RELEASE.jar spring-context-support-5
+ .0.2.RELEASE.jar spring-plugin-metadata-1.2.0.RELEASE.jar springfox-s
+ chema-2.5.0.jar spring-boot-starter-tomcat-1.5.2.RELEASE.jar
+Main-Class: com.gsafety.hikaru.application.Application
+
