+++ hikaru-api/pom.xml
@@ -26,11 +26,5 @@
             <version>1.0-SNAPSHOT</version>
         </dependency>
 
-        <dependency>
-            <groupId>com.gsafety.hikaru</groupId>
-            <artifactId>hikaru-common</artifactId>
-            <version>1.0-SNAPSHOT</version>
-        </dependency>
-
     </dependencies>
 </project>
