+++ hikaru-application/pom.xml
@@ -61,15 +61,4 @@
 
     </dependencies>
 
-    <build>
-        <finalName>hikaru-server</finalName>
-        <plugins>
-            <plugin>
-                <groupId>org.springframework.boot</groupId>
-                <artifactId>spring-boot-maven-plugin</artifactId>
-            </plugin>
-        </plugins>
-    </build>
-
-
 </project>
