+++ hikaru-application/src/main/resources/application.properties
@@ -10,3 +10,19 @@ language=zh_CN
 #spring.kafka.consumer.group-id=testGroup
 
 
+#spring.cloud.consul.port=8500
+#spring.cloud.consul.host=localhost
+#spring.cloud.consul.enabled=true
+#spring.cloud.consul.discovery.enabled=true
+#spring.cloud.consul.discovery.instance-id=${spring.application.name}-${server.address}
+#spring.cloud.consul.discovery.service-name=${spring.application.name}
+#spring.cloud.consul.discovery.health-check-url=http://${server.address}:${server.port}/${server.servlet.context-path}/health
+#spring.cloud.consul.discovery.hostname=${server.address}
+#spring.cloud.consul.discovery.port=${server.port}
+#spring.cloud.consul.discovery.health-check-interval=10s
+#spring.cloud.consul.discovery.tags=dev
+#spring.cloud.consul.config.enabled=true
+#spring.cloud.consul.config.format=PROPERTIES
+#spring.cloud.consul.config.data-key=configuration
+
+
