+++ hikaru-common/pom.xml
@@ -8,9 +8,6 @@
         <version>1.0-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
-    <modules>
-        <module></module>
-    </modules>
     <artifactId>hikaru-common</artifactId>
 
     <dependencies>
