+++ hikaru-core/src/test/java/zjj/LoopTest.java
@@ -50,7 +50,7 @@ public class LoopTest {
         loop.setCircle(circle);
     }
 
-    @Test
+//    @Test
     public void test() {
 
         curve = loop.calculateCurve();
