+++ hikaru-core/src/test/java/zjj/StreamTest.java
@@ -36,7 +36,7 @@ public class StreamTest {
         log.log(firstElements);
     }
 
-    @Test
+//    @Test
     public void test() throws IOException {
         Path path = Paths.get("D:\\WorkSpace\\Test\\Personal\\Java\\kafka-server\\kafka-core\\src\\main\\resources\\properties\\db.properties");
         String contents = new String(Files.readAllBytes(path), StandardCharsets.UTF_8);
