+++ hikaru-model/pom.xml
@@ -13,12 +13,6 @@
     <dependencies>
         <dependency>
             <groupId>com.gsafety.hikaru</groupId>
-            <artifactId>hikaru-common</artifactId>
-            <version>1.0-SNAPSHOT</version>
-        </dependency>
-
-        <dependency>
-            <groupId>com.gsafety.hikaru</groupId>
             <artifactId>hikaru-core</artifactId>
             <version>1.0-SNAPSHOT</version>
         </dependency>
