+++ hikaru-service/pom.xml
@@ -23,5 +23,11 @@
             <artifactId>hikaru-model</artifactId>
             <version>1.0-SNAPSHOT</version>
         </dependency>
+
+        <dependency>
+            <groupId>com.gsafety.hikaru</groupId>
+            <artifactId>hikaru-common</artifactId>
+            <version>1.0-SNAPSHOT</version>
+        </dependency>
     </dependencies>
 </project>
